refactor: extract shared Friendly Captcha widget component
