+++ lib/plausible_web/live/register_form.ex
@@ -131,85 +131,16 @@ defmodule PlausibleWeb.Live.RegisterForm do
         </div>
 
         <%= if PlausibleWeb.Captcha.enabled?() do %>
-          <div>
-            <div
-              phx-update="ignore"
-              id="frc-captcha-placeholder"
-              class="frc-captcha hidden"
-              data-sitekey={PlausibleWeb.Captcha.sitekey()}
-              data-start="auto"
-            >
-            </div>
-            <p :if={@captcha_error} class="text-xs text-red-500 mt-2">
-              {@captcha_error}
-            </p>
-            <p class="text-xs text-gray-500 dark:text-gray-400">
-              This site is protected by
-              <.styled_link href="https://friendlycaptcha.com" new_tab={true}>
-                Friendly Captcha
-              </.styled_link>
-            </p>
-            <script
-              phx-update="ignore"
-              id="frc-captcha-script"
-              type="module"
-              src="https://cdn.jsdelivr.net/npm/@friendlycaptcha/sdk@1/site.min.js"
-              async
-              defer
-            >
-            </script>
-            <script
-              phx-update="ignore"
-              id="frc-captcha-script-compat"
-              nomodule
-              src="https://cdn.jsdelivr.net/npm/@friendlycaptcha/sdk@1/site.compat.min.js"
-              async
-              defer
-            >
-            </script>
-            <script phx-update="ignore" id="frc-captcha-reveal">
-              (function () {
-                var SHOW_AFTER_LONG_WAIT_MS = 5000;
-                var el = document.getElementById("frc-captcha-placeholder");
-                if (!el) return;
-
-                // Match the widget to the app's resolved light/dark theme. This runs
-                // before the (deferred) SDK initializes, so the widget picks up the
-                // right theme from the start, and the observer keeps it in sync.
-                function applyTheme() {
-                  el.dataset.theme =
-                    document.documentElement.classList.contains("dark") ? "dark" : "light";
-                }
-                applyTheme();
-                new MutationObserver(applyTheme).observe(document.documentElement, {
-                  attributes: true,
-                  attributeFilter: ["class"]
-                });
-
-                function show() { el.classList.remove("hidden"); }
-                var timeout;
-                // Friendly Captcha carries the event payload on `e.detail` (not `e`).
-                el.addEventListener("frc:widget.statechange", function (e) {
-                  var d = e.detail || {};
-                  // Interactive mode means the user must click to solve: reveal the widget.
-                  if (d.mode === "interactive") { show(); }
-                  // Reveal if solving takes unusually long, then stop waiting once done.
-                  if (d.state === "requesting") {
-                    clearTimeout(timeout);
-                    timeout = setTimeout(show, SHOW_AFTER_LONG_WAIT_MS);
-                  } else if (d.state === "completed") {
-                    clearTimeout(timeout);
-                  }
-                  // Reveal on error or expiry so the user can recover.
-                  if (d.state === "error" || d.state === "expired") { show(); }
-                  // Enable the submit button only once we hold a valid solution.
-                  window.dispatchEvent(new Event(
-                    d.state === "completed" ? "frc-captcha-ready" : "frc-captcha-reset"
-                  ));
-                });
-              })();
-            </script>
-          </div>
+          <PlausibleWeb.Components.Captcha.widget live?={true} error={@captcha_error}>
+            <:attribution>
+              <p class="text-xs text-gray-500 dark:text-gray-400">
+                This site is protected by
+                <.styled_link href="https://friendlycaptcha.com" new_tab={true}>
+                  Friendly Captcha
+                </.styled_link>
+              </p>
+            </:attribution>
+          </PlausibleWeb.Components.Captcha.widget>
         <% end %>
 
         <div class="flex flex-col gap-y-4">
@@ -224,11 +155,10 @@ defmodule PlausibleWeb.Live.RegisterForm do
             type="submit"
             class="w-full"
             mt?={false}
-            data-disable-submit={to_string(@disable_submit)}
             x-data={"{ captchaReady: #{not PlausibleWeb.Captcha.enabled?()} }"}
             x-on:frc-captcha-ready.window="captchaReady = true"
             x-on:frc-captcha-reset.window="captchaReady = false"
-            x-bind:disabled="!captchaReady || $el.dataset.disableSubmit === 'true'"
+            x-bind:disabled={"!captchaReady || #{@disable_submit}"}
           >
             {submit_text}
           </.button>
