+++ lib/plausible_web/templates/auth/password_reset_request_form.html.heex
@@ -25,86 +25,25 @@
     </div>
 
     <%= if PlausibleWeb.Captcha.enabled?() do %>
-      <div>
-        <div
-          id="frc-captcha-placeholder"
-          class="frc-captcha hidden"
-          data-sitekey={PlausibleWeb.Captcha.sitekey()}
-          data-start="auto"
-        >
-        </div>
-        <p :if={assigns[:captcha_error]} class="text-xs text-red-500 mt-2">
-          {@captcha_error}
-        </p>
-        <p class="text-xs text-center text-gray-500 dark:text-gray-400">
-          This site is protected by
-          <.styled_link href="https://friendlycaptcha.com" new_tab={true}>
-            Friendly Captcha
-          </.styled_link>
-          and its
-          <.styled_link href="https://friendlycaptcha.com/privacy/user-protection/" new_tab={true}>
-            Privacy Policy
-          </.styled_link>
-          and
-          <.styled_link href="https://friendlycaptcha.com/legal/terms/" new_tab={true}>
-            Terms of Use
-          </.styled_link>
-          apply.
-        </p>
-        <script
-          type="module"
-          src="https://cdn.jsdelivr.net/npm/@friendlycaptcha/sdk@1/site.min.js"
-          async
-          defer
-        >
-        </script>
-        <script
-          nomodule
-          src="https://cdn.jsdelivr.net/npm/@friendlycaptcha/sdk@1/site.compat.min.js"
-          async
-          defer
-        >
-        </script>
-        <script>
-          (function () {
-            var SHOW_AFTER_LONG_WAIT_MS = 5000;
-            var el = document.getElementById("frc-captcha-placeholder");
-            if (!el) return;
-
-            // Match the widget to the app's resolved light/dark theme. This runs
-            // before the (deferred) SDK initializes, so the widget picks up the
-            // right theme from the start, and the observer keeps it in sync.
-            function applyTheme() {
-              el.dataset.theme =
-                document.documentElement.classList.contains("dark") ? "dark" : "light";
-            }
-            applyTheme();
-            new MutationObserver(applyTheme).observe(document.documentElement, {
-              attributes: true,
-              attributeFilter: ["class"]
-            });
-
-            function show() { el.classList.remove("hidden"); }
-            var timeout;
-            // Friendly Captcha carries the event payload on `e.detail` (not `e`).
-            el.addEventListener("frc:widget.statechange", function (e) {
-              var d = e.detail || {};
-              if (d.mode === "interactive") { show(); }
-              if (d.state === "requesting") {
-                clearTimeout(timeout);
-                timeout = setTimeout(show, SHOW_AFTER_LONG_WAIT_MS);
-              } else if (d.state === "completed") {
-                clearTimeout(timeout);
-              }
-              if (d.state === "error" || d.state === "expired") { show(); }
-              // Enable the submit button only once we hold a valid solution.
-              window.dispatchEvent(new Event(
-                d.state === "completed" ? "frc-captcha-ready" : "frc-captcha-reset"
-              ));
-            });
-          })();
-        </script>
-      </div>
+      <PlausibleWeb.Components.Captcha.widget error={assigns[:captcha_error]}>
+        <:attribution>
+          <p class="text-xs text-center text-gray-500 dark:text-gray-400">
+            This site is protected by
+            <.styled_link href="https://friendlycaptcha.com" new_tab={true}>
+              Friendly Captcha
+            </.styled_link>
+            and its
+            <.styled_link href="https://friendlycaptcha.com/privacy/user-protection/" new_tab={true}>
+              Privacy Policy
+            </.styled_link>
+            and
+            <.styled_link href="https://friendlycaptcha.com/legal/terms/" new_tab={true}>
+              Terms of Use
+            </.styled_link>
+            apply.
+          </p>
+        </:attribution>
+      </PlausibleWeb.Components.Captcha.widget>
     <% end %>
 
     <div class="flex flex-col gap-y-4">
